+++ src/app/routes/ui-demo/list-form/list-form.component.ts
@@ -58,7 +58,7 @@ export class ListFormComponent extends SipModal {
 
     @SipNgDestroy()
     private _destroy() {
-        this.$logger.debug('_destroy');
+        this.$logger.debug('_destroy list from');
     }
 
     save() {