+++ projects/sip-alain/CHANGELOG.md
@@ -7,6 +7,7 @@
 * 优化：ng生命周期事件
 * 添加: SipPrepareData && SipInit
 * 添加：$onDestroy事件
+* 优化：ng event
 
 ## 0.6.0